pyproject.toml
[project]
name = "llm-latent-dspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "dspy-ai>=2.4.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.27.0",
    "groq>=0.4.0",
    "datasets>=2.14.0",
    "numpy>=1.24.0",
    "huggingface-hub>=0.34.3",
]
